+++ packages/headless/src/features/insight-interface/insight-interface-state.ts
@@ -1,0 +1,16 @@
+import {InsightInterface} from '../../api/service/insight/get-interface/get-interface-response';
+import {InsightAPIErrorStatusResponse} from '../../api/service/insight/insight-api-client';
+
+export const getInsightInterfaceInitialState = (): InsightInterfaceState => ({
+  loading: false,
+  config: undefined,
+});
+
+export interface InsightInterfaceState {
+  loading: boolean;
+  config?: {
+    contextFields: Record<string, string | number | string[]>;
+    interface?: InsightInterface;
+  };
+  error?: InsightAPIErrorStatusResponse;
+}